docs: ✏️ improve swagger service description
