+++ src/common/utils/consts.ts
@@ -1,0 +1,15 @@
+export const swaggerTitle = 'Polymesh REST API';
+
+export const swaggerDescription = [
+  'The Polymesh REST API provides a developer friendly interface with the Polymesh blockchain',
+  'Polymesh is an institutional-grade permissioned blockchain built specifically for regulated assets.',
+  '',
+  'The API allows you to perform various operations such as:',
+  '',
+  '- Query and manage identity data',
+  '- Interact with corporate actions (e.g., dividends, capital distributions, and other events)',
+  '- Create and manage security tokens',
+  '- Manage asset ownership and transfers',
+  '',
+  'With this API developers can build applications and integrations for the Polymesh chain in the programming language of their choice',
+].join('\n');
